# apps/api/src/modules/leaderboards/leaderboards.controller.ts
import { Controller, Get, Headers, Query } from '@nestjs/common';
import { ApiBearerAuth, ApiOkResponse, ApiTags } from '@nestjs/swagger';
import { DifficultyQueryDto } from '../pvp/pvp.dto';
import { LeaderboardDto } from './leaderboards.dto';
import { LeaderboardsService } from './leaderboards.service';

@ApiTags('leaderboards')
@ApiBearerAuth()
@Controller('leaderboards')
export class LeaderboardsController {
  constructor(private readonly service: LeaderboardsService) {}
  @Get('pvp')
  @ApiOkResponse({ type: LeaderboardDto })
  pvp(
    @Headers('authorization') authorization: string | undefined,
    @Query() query: DifficultyQueryDto,
  ) {
    return this.service.pvp(authorization, query.difficulty);
  }
  @Get('class')
  @ApiOkResponse({ type: LeaderboardDto })
  class(@Headers('authorization') authorization?: string) {
    return this.service.class(authorization);
  }
}