# packages/contracts/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tryout/packages/current": {
      "get": {
        "operationId": "TryoutController_current",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentTryoutDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tryout"
        ]
      }
    },
    "/api/v1/tryout/attempts": {
      "post": {
        "operationId": "TryoutController_start",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTryoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TryoutAttemptDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tryout"
        ]
      }
    },
    "/api/v1/tryout/attempts/{attemptId}": {
      "get": {
        "operationId": "TryoutController_attempt",
        "parameters": [
          {
            "name": "attemptId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TryoutAttemptDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tryout"
        ]
      }
    },
    "/api/v1/tryout/attempts/{attemptId}/answers/{questionInstanceId}": {
      "patch": {
        "operationId": "TryoutController_saveAnswer",
        "parameters": [
          {
            "name": "attemptId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "questionInstanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDrillAnswerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedAnswerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tryout"
        ]
      }
    },
    "/api/v1/tryout/attempts/{attemptId}/submit": {
      "post": {
        "operationId": "TryoutController_submit",
        "parameters": [
          {
            "name": "attemptId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TryoutSubmitDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tryout"
        ]
      }
    },
    "/api/v1/tryout/attempts/{attemptId}/result": {
      "get": {
        "operationId": "TryoutController_result",
        "parameters": [
          {
            "name": "attemptId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TryoutResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tryout"
        ]
      }
    },
...
    "/api/v1/leaderboards/pvp": {
      "get": {
        "operationId": "LeaderboardsController_pvp",
        "parameters": [
          {
            "name": "difficulty",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "easy",
                "medium",
                "hard"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "leaderboards"
        ]
      }
    },
    "/api/v1/leaderboards/class": {
      "get": {
        "operationId": "LeaderboardsController_class",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "leaderboards"
        ]
      }
    }
  },
  "info": {
    "title": "TKA Mathematics SMP API",
    "description": "REST contract for the TKA Mathematics SMP platform.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "text"
        ]
      },
      "DrillQuestionDto": {
        "type": "object",
        "properties": {
          "questionInstanceId": {
            "type": "string",
            "format": "uuid"
          },
          "stem": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OptionDto"
            }
          },
          "selectedOptionId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "questionInstanceId",
          "stem",
          "options",
          "selectedOptionId"
        ]
      },
...
      "SaveDrillAnswerDto": {
        "type": "object",
        "properties": {
          "optionId": {
            "type": "string",
            "nullable": true,
            "description": "A-D for demo; null clears the answer."
          }
        },
        "required": [
          "optionId"
        ]
      },
      "SavedAnswerDto": {
        "type": "object",
        "properties": {
          "questionInstanceId": {
            "type": "string",
            "format": "uuid"
          },
          "selectedOptionId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "questionInstanceId",
          "selectedOptionId"
        ]
      },
...
      "CurrentTryoutDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "releaseAt": {
            "type": "string",
            "format": "date-time"
          },
          "state": {
            "type": "string",
            "enum": [
              "unavailable",
              "open",
              "inProgress",
              "waitingIrt",
              "resultReady"
            ]
          },
          "eligible": {
            "type": "boolean"
          },
          "attemptId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "questionCount": {
            "type": "number",
            "nullable": true
          },
          "durationSeconds": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "state"
        ]
      },
      "StartTryoutDto": {
        "type": "object",
        "properties": {
          "packageId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "packageId"
        ]
      },
      "TryoutAttemptDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "packageId": {
            "type": "string",
            "format": "uuid"
          },
          "packageTitle": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "inProgress",
              "submitted"
            ]
          },
          "deadlineAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DrillQuestionDto"
            }
          }
        },
        "required": [
          "id",
          "packageId",
          "packageTitle",
          "status",
          "deadlineAt",
          "questions"
        ]
      },
      "TryoutSubmitDto": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "waitingIrt"
            ]
          }
        },
        "required": [
          "state"
        ]
      },
      "TryoutReviewedQuestionDto": {
        "type": "object",
        "properties": {
          "questionInstanceId": {
            "type": "string",
            "format": "uuid"
          },
          "stem": {
            "type": "string"
          },
          "selectedOptionId": {
            "type": "string",
            "nullable": true
          },
          "correctOptionId": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          }
        },
        "required": [
          "questionInstanceId",
          "stem",
          "selectedOptionId",
          "correctOptionId",
          "explanation"
        ]
      },
      "TryoutResultDto": {
        "type": "object",
        "properties": {
          "attemptId": {
            "type": "string",
            "format": "uuid"
          },
          "packageTitle": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "correctCount": {
            "type": "number"
          },
          "questionCount": {
            "type": "number"
          },
          "explanation": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TryoutReviewedQuestionDto"
            }
          }
        },
        "required": [
          "attemptId",
          "packageTitle",
          "score",
          "correctCount",
          "questionCount",
          "explanation"
        ]
      },
...
      "LeaderboardPeriodDto": {
        "type": "object",
        "properties": {
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          }
        },
        "required": [
          "startsAt",
          "endsAt",
          "timezone"
        ]
      },
      "LeaderboardEntryDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "points": {
            "type": "number"
          },
          "rank": {
            "type": "number"
          }
        },
        "required": [
          "studentId",
          "displayName",
          "points",
          "rank"
        ]
      },
      "LeaderboardDto": {
        "type": "object",
        "properties": {
          "policyPending": {
            "type": "boolean"
          },
          "reasonCode": {
            "type": "string",
            "nullable": true
          },
          "className": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "enum": [
              "points",
              "xp"
            ]
          },
          "period": {
            "$ref": "#/components/schemas/LeaderboardPeriodDto"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeaderboardEntryDto"
            }
          },
          "ownEntry": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LeaderboardEntryDto"
              }
            ]
          }
        },
        "required": [
          "policyPending",
          "reasonCode",
          "className",
          "unit",
          "period",
          "updatedAt",
          "entries",
          "ownEntry"
        ]
      }
    }
  }
}